Add yaml/deploy-jenkins.yaml
Test and find warning. Detail see https://github.com/jenkinsci/docker/issues/397

The Jenkins running on default namespace has the same warning
It is triggered when tring to install Nexus-Jenkins-plugin
$ kubectl logs jenkins-2377375696-8qevu | grep newHostInfo -C 200 | more

Jenkins works with this warning.
$ kubectl exec sonarqube-2950109087-nwgpr --namespace=white  -it curl http://10.36.0.6:8080
sonarqube-2950109087-nwgpr here is a pod running in the same k8s cluster.

Note: the Password of admin user is generated and is saved in the file.
/var/jenkins_home/secrets/initialAdminPassword