Specify the image version
Don not use the default version 'latest' because sometimes it is not stable.
Also uniform the depolyment name and label name. Start with namespace name to
make sure even some wrong operation, e.g. without specify the namespace in kubectl
command, will not influcence other application running in default namespace.

Test and find it works smoothly without any issue in the whole process
of deployment using kubectl attach command.

It works.
$ kubectl exec sonarqube-2950109087-nwgpr --namespace=white  -it curl http://10.32.0.3:9000
Here sonarqube-2950109087-nwgpr is a pod in the same k8s cluster. (It is the old pod of
sonarqube and will be deleted to save resource)