Add playbook.yaml
It works and find Ansible cannot guarantee the idempotency for kubernetes.
That means we need delete every thing if they exist before we start to
deploy them automatically, service, deployments and namespace at present
in the reverse order of creating them.

In the deploy yaml files namespace is still hardcode