Donot let some failed handler stop executing other handlers
Handlers will run only once after all of the tasks complete.
But handlers always run in the same order they are defined, not
in the order listed in the notify-statement. So if some handler
defined in a previous location failed the handlers defined
after it will not execute even they are notified in advance of
the falied one. This makes it hard to debug. Plus handlers
have not dependency relation at present, so do not let some failed
handler stop executing other handlers.