Make sure namespace is created before deployment starts
Some times the ansible task to create namespace in k8s says
it is done but in fact in k8s the process is still going on.
In this case if next task starts quickly enough before the k8s
namespace is got ready, error will is reported and
ansible-playbook will stop there.

Pause 3 seconds to avoid this situation happens.